Continuing unit tests
Unit tests working for :
 - PWM clutch outputs
 - telemetry sender (with functional Python receiver script)
 - potentiometer sampling at 1KHz

+++ PrometheUS_F446ZE/Core/PrometheUS/Sensing/GenericAdcBurstSampler.hpp
@@ -15,7 +15,7 @@ public:
   virtual void convertAll(float (&outValues)[FINGER_COUNT]) const = 0;
 
 protected:
-  uint32_t m_rawAdcValueArray[FINGER_COUNT] = { 0 };
+  uint16_t m_rawAdcValueArray[FINGER_COUNT] = { 0 };
 
 private:
   ADC_HandleTypeDef* m_hadc = nullptr;
